+++ lib/extension/dialog_assets.dart
@@ -1,0 +1,19 @@
+/*
+ * @Author: lipeng 1162423147@qq.com
+ * @Date: 2023-09-29 10:44:56
+ * @LastEditors: lipeng 1162423147@qq.com
+ * @LastEditTime: 2023-09-29 10:46:33
+ * @FilePath: /phoenix_dialog/lib/extension/dialog_assets.dart
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+ */
+import 'package:phoenix_base/phoenix.dart';
+
+extension DialogAssets on BaseAsset {
+  static const String iconUnSelect = "images/icon_unselect.png";
+  static const String iconPickerClose = 'images/icon_guanbi.png';
+  static const String iconAlert = "icons/icon_alert.png";
+  static const String iconWarning = "icons/icon_warning.png";
+  static const String iconSuccess = "icons/icon_success.png";
+  static const String iconMultiSelected = "images/icon_multi_selected.png";
+  static const String iconSingleSelected = "images/icon_single_selected.png";
+}