+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+<!--
+ * @Author: lipeng 1162423147@qq.com
+ * @Date: 2023-09-24 11:10:50
+ * @LastEditors: lipeng 1162423147@qq.com
+ * @LastEditTime: 2023-09-29 10:50:35
+ * @FilePath: /phoenix_dialog/CHANGELOG.md
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+-->
+## 1.0.0
 
-* TODO: Describe initial release.
+* [新建] 企业级基础组件:Dialog
+
+## 1.0.1
+
+* [修改] resource path